pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
]
